requirements.txt
# Core Data Processing
pandas>=2.1.0
pyarrow>=14.0.0
requests>=2.31.0

# Data Quality & Validation
great-expectations>=0.18.0
pandera>=0.17.0

# Delta Lake (for local development)
deltalake>=0.15.0

# Logging & Monitoring
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
pylint>=3.0.0
mypy>=1.7.0

# Utilities
tqdm>=4.66.0

# Visualization (for notebooks)
matplotlib>=3.8.0
seaborn>=0.13.0

# Optional: Para uso com PySpark (descomentar se necessário)
# pyspark>=3.5.0
